Create separate modules for aperture shapes
