+++ photutils/aperture/__init__.py
@@ -4,6 +4,9 @@ This subpackage contains modules and packages for identifying sources in
 an astronomical image.
 """
 
-from .core import *
 from .aperture_funcs import *
+from .circle import *
+from .core import *
+from .ellipse import *
 from .mask_data import *
+from .rectangle import *
