+++ photutils/aperture/core.py
@@ -1,16 +1,12 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
-"""Functions for performing aperture photometry on 2-D arrays."""
-
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 import abc
 import copy
-import math
 import warnings
 from collections import OrderedDict
 
 import numpy as np
-from astropy.coordinates import SkyCoord
 from astropy.extern import six
 from astropy.io import fits
 from astropy.nddata import support_nddata
@@ -19,28 +15,9 @@ import astropy.units as u
 from astropy.utils.exceptions import AstropyUserWarning
 from astropy.utils.misc import InheritDocstrings
 from astropy.wcs import WCS
-from astropy.wcs.utils import skycoord_to_pixel
-
-from .aperture_funcs import (do_circular_photometry, get_circular_fractions,
-                             do_elliptical_photometry,
-                             get_elliptical_fractions,
-                             do_rectangular_photometry,
-                             get_rectangular_fractions)
-from ..utils.wcs_helpers import (skycoord_to_pixel_scale_angle, assert_angle,
-                                 assert_angle_or_pixel)
 
 
-skycoord_to_pixel_mode = 'all'
-
-
-__all__ = ['Aperture', 'SkyAperture', 'PixelAperture',
-           'SkyCircularAperture', 'CircularAperture',
-           'SkyCircularAnnulus', 'CircularAnnulus',
-           'SkyEllipticalAperture', 'EllipticalAperture',
-           'SkyEllipticalAnnulus', 'EllipticalAnnulus',
-           'RectangularAperture', 'RectangularAnnulus',
-           'SkyRectangularAperture', 'SkyRectangularAnnulus',
-           'aperture_photometry']
+__all__ = ['Aperture', 'SkyAperture', 'PixelAperture', 'aperture_photometry']
 
 
 def _get_version_info():
@@ -312,987 +289,6 @@ class PixelAperture(Aperture):
         area : float
             Area of aperture.
         """
-
-
-class SkyCircularAperture(SkyAperture):
-    """
-    Circular aperture(s), defined in sky coordinates.
-
-    Parameters
-    ----------
-    positions : `~astropy.coordinates.SkyCoord`
-        Celestial coordinates of the aperture center(s). This can be either
-        scalar coordinates or an array of coordinates.
-    r : `~astropy.units.Quantity`
-        The radius of the aperture(s), either in angular or pixel units.
-    """
-
-    def __init__(self, positions, r):
-
-        if isinstance(positions, SkyCoord):
-            self.positions = positions
-        else:
-            raise TypeError("positions should be a SkyCoord instance")
-
-        assert_angle_or_pixel('r', r)
-        self.r = r
-
-    def to_pixel(self, wcs):
-        """
-        Return a CircularAperture instance in pixel coordinates.
-        """
-
-        x, y = skycoord_to_pixel(self.positions, wcs,
-                                 mode=skycoord_to_pixel_mode)
-
-        if self.r.unit.physical_type == 'angle':
-            central_pos = SkyCoord([wcs.wcs.crval], frame=self.positions.name,
-                                   unit=wcs.wcs.cunit)
-            xc, yc, scale, angle = skycoord_to_pixel_scale_angle(central_pos,
-                                                                 wcs)
-            r = (scale * self.r).to(u.pixel).value
-        else:  # pixel
-            r = self.r.value
-
-        pixel_positions = np.array([x, y]).transpose()
-
-        return CircularAperture(pixel_positions, r)
-
-
-class CircularAperture(PixelAperture):
-    """
-    Circular aperture(s), defined in pixel coordinates.
-
-    Parameters
-    ----------
-    positions : tuple, list, array, or `~astropy.units.Quantity`
-        Pixel coordinates of the aperture center(s), either as a single
-        ``(x, y)`` tuple, a list of ``(x, y)`` tuples, an ``Nx2`` or
-        ``2xN`` `~numpy.ndarray`, or an ``Nx2`` or ``2xN``
-        `~astropy.units.Quantity` in units of pixels.  A ``2x2``
-        `~numpy.ndarray` or `~astropy.units.Quantity` is interpreted as
-        ``Nx2``, i.e. two rows of (x, y) coordinates.
-    r : float
-        The radius of the aperture(s), in pixels.
-
-    Raises
-    ------
-    ValueError : `ValueError`
-        If the radius is negative.
-    """
-
-    def __init__(self, positions, r):
-
-        try:
-            self.r = float(r)
-        except TypeError:
-            raise TypeError('r must be numeric, received {0}'.format(type(r)))
-
-        if r < 0:
-            raise ValueError('r must be non-negative')
-
-        self.positions = _sanitize_pixel_positions(positions)
-
-    def area(self):
-        return math.pi * self.r ** 2
-
-    def plot(self, origin=(0, 0), source_id=None, ax=None, fill=False,
-             **kwargs):
-
-        import matplotlib.patches as mpatches
-
-        plot_positions, ax, kwargs = self._prepare_plot(
-            origin, source_id, ax, fill, **kwargs)
-
-        for position in plot_positions:
-            patch = mpatches.Circle(position, self.r, **kwargs)
-            ax.add_patch(patch)
-
-    def do_photometry(self, data, error=None, pixelwise_error=True,
-                      method='exact', subpixels=5):
-
-        if method not in ('center', 'subpixel', 'exact'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        flux = do_circular_photometry(data, self.positions,
-                                      self.r, error=error,
-                                      pixelwise_error=pixelwise_error,
-                                      method=method,
-                                      subpixels=subpixels)
-        return flux
-
-    def get_fractions(self, data, method='exact', subpixels=5):
-
-        if method not in ('center', 'subpixel', 'exact'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        fractions = get_circular_fractions(data, self.positions,
-                                           self.r, method=method,
-                                           subpixels=subpixels)
-        return fractions
-
-
-class SkyCircularAnnulus(SkyAperture):
-    """
-    Circular annulus aperture(s), defined in sky coordinates.
-
-    Parameters
-    ----------
-
-    positions : `~astropy.coordinates.SkyCoord`
-        Celestial coordinates of the aperture center(s). This can be either
-        scalar coordinates or an array of coordinates.
-
-    r_in : `~astropy.units.Quantity`
-        The inner radius of the annulus, either in angular or pixel units.
-
-    r_out : `~astropy.units.Quantity`
-        The outer radius of the annulus, either in angular or pixel units.
-    """
-
-    def __init__(self, positions, r_in, r_out):
-
-        if isinstance(positions, SkyCoord):
-            self.positions = positions
-        else:
-            raise TypeError("positions should be a SkyCoord instance")
-
-        assert_angle_or_pixel('r_in', r_in)
-        assert_angle_or_pixel('r_out', r_out)
-
-        if r_in.unit.physical_type != r_out.unit.physical_type:
-            raise ValueError("r_in and r_out should either both be angles "
-                             "or in pixels")
-
-        self.r_in = r_in
-        self.r_out = r_out
-
-    def to_pixel(self, wcs):
-        """
-        Return a CircularAnnulus instance in pixel coordinates.
-        """
-
-        x, y = skycoord_to_pixel(self.positions, wcs,
-                                 mode=skycoord_to_pixel_mode)
-        if self.r_in.unit.physical_type == 'angle':
-            central_pos = SkyCoord([wcs.wcs.crval], frame=self.positions.name,
-                                   unit=wcs.wcs.cunit)
-            xc, yc, scale, angle = skycoord_to_pixel_scale_angle(central_pos,
-                                                                 wcs)
-            r_in = (scale * self.r_in).to(u.pixel).value
-            r_out = (scale * self.r_out).to(u.pixel).value
-        else:  # pixel
-            r_in = self.r_in.value
-            r_out = self.r_out.value
-
-        pixel_positions = np.array([x, y]).transpose()
-
-        return CircularAnnulus(pixel_positions, r_in, r_out)
-
-
-class CircularAnnulus(PixelAperture):
-    """
-    Circular annulus aperture(s), defined in pixel coordinates.
-
-    Parameters
-    ----------
-    positions : tuple, list, array, or `~astropy.units.Quantity`
-        Pixel coordinates of the aperture center(s), either as a single
-        ``(x, y)`` tuple, a list of ``(x, y)`` tuples, an ``Nx2`` or
-        ``2xN`` `~numpy.ndarray`, or an ``Nx2`` or ``2xN``
-        `~astropy.units.Quantity` in units of pixels.  A ``2x2``
-        `~numpy.ndarray` or `~astropy.units.Quantity` is interpreted as
-        ``Nx2``, i.e. two rows of (x, y) coordinates.
-    r_in : float
-        The inner radius of the annulus.
-    r_out : float
-        The outer radius of the annulus.
-
-    Raises
-    ------
-    ValueError : `ValueError`
-        If inner radius (``r_in``) is greater than outer radius (``r_out``).
-    ValueError : `ValueError`
-        If inner radius is negative.
-    """
-
-    def __init__(self, positions, r_in, r_out):
-        try:
-            self.r_in = r_in
-            self.r_out = r_out
-        except TypeError:
-            raise TypeError("'r_in' and 'r_out' must be numeric, received "
-                            "{0} and {1}".format((type(r_in), type(r_out))))
-
-        if not (r_out > r_in):
-            raise ValueError('r_out must be greater than r_in')
-        if r_in < 0:
-            raise ValueError('r_in must be non-negative')
-
-        self.positions = _sanitize_pixel_positions(positions)
-
-    def area(self):
-        return math.pi * (self.r_out ** 2 - self.r_in ** 2)
-
-    def do_photometry(self, data, error=None, pixelwise_error=True,
-                      method='exact', subpixels=5):
-
-        if method not in ('center', 'subpixel', 'exact'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        flux = do_circular_photometry(data, self.positions,
-                                      self.r_out, error=error,
-                                      pixelwise_error=pixelwise_error,
-                                      method=method,
-                                      subpixels=subpixels,
-                                      r_in=self.r_in)
-
-        return flux
-
-    def get_fractions(self, data, method='exact', subpixels=5):
-
-        if method not in ('center', 'subpixel', 'exact'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        fractions = get_circular_fractions(data, self.positions,
-                                           self.r_out, method=method,
-                                           subpixels=subpixels,
-                                           r_in=self.r_in)
-
-        return fractions
-
-    def plot(self, origin=(0, 0), source_id=None, ax=None, fill=False,
-             **kwargs):
-
-        import matplotlib.patches as mpatches
-
-        plot_positions, ax, kwargs = self._prepare_plot(
-            origin, source_id, ax, fill, **kwargs)
-
-        resolution = 20
-        for position in plot_positions:
-            patch_inner = mpatches.CirclePolygon(position, self.r_in,
-                                                 resolution=resolution)
-            patch_outer = mpatches.CirclePolygon(position, self.r_out,
-                                                 resolution=resolution)
-            path = _make_annulus_path(patch_inner, patch_outer)
-            patch = mpatches.PathPatch(path, **kwargs)
-            ax.add_patch(patch)
-
-
-class SkyEllipticalAperture(SkyAperture):
-    """
-    Elliptical aperture(s), defined in sky coordinates.
-
-    Parameters
-    ----------
-    positions : `~astropy.coordinates.SkyCoord`
-        Celestial coordinates of the aperture center(s). This can be either
-        scalar coordinates or an array of coordinates.
-    a : `~astropy.units.Quantity`
-        The semimajor axis, either in angular or pixel units.
-    b : `~astropy.units.Quantity`
-        The semiminor axis, either in angular or pixel units.
-    theta : `~astropy.units.Quantity`
-        The position angle of the semimajor axis (counterclockwise), either
-        in angular or pixel units.
-    """
-
-    def __init__(self, positions, a, b, theta):
-
-        if isinstance(positions, SkyCoord):
-            self.positions = positions
-        else:
-            raise TypeError("positions should be a SkyCoord instance")
-
-        assert_angle_or_pixel('a', a)
-        assert_angle_or_pixel('b', b)
-        assert_angle('theta', theta)
-
-        if a.unit.physical_type != b.unit.physical_type:
-            raise ValueError("a and b should either both be angles "
-                             "or in pixels")
-
-        self.a = a
-        self.b = b
-        self.theta = theta
-
-    def to_pixel(self, wcs):
-        """
-        Return a EllipticalAperture instance in pixel coordinates.
-        """
-
-        x, y = skycoord_to_pixel(self.positions, wcs,
-                                 mode=skycoord_to_pixel_mode)
-        central_pos = SkyCoord([wcs.wcs.crval], frame=self.positions.name,
-                               unit=wcs.wcs.cunit)
-        xc, yc, scale, angle = skycoord_to_pixel_scale_angle(central_pos, wcs)
-
-        if self.a.unit.physical_type == 'angle':
-            a = (scale * self.a).to(u.pixel).value
-            b = (scale * self.b).to(u.pixel).value
-        else:  # pixel
-            a = self.a.value
-            b = self.b.value
-
-        theta = (angle + self.theta).to(u.radian).value
-
-        pixel_positions = np.array([x, y]).transpose()
-
-        return EllipticalAperture(pixel_positions, a, b, theta)
-
-
-class EllipticalAperture(PixelAperture):
-    """
-    Elliptical aperture(s), defined in pixel coordinates.
-
-    Parameters
-    ----------
-    positions : tuple, list, array, or `~astropy.units.Quantity`
-        Pixel coordinates of the aperture center(s), either as a single
-        ``(x, y)`` tuple, a list of ``(x, y)`` tuples, an ``Nx2`` or
-        ``2xN`` `~numpy.ndarray`, or an ``Nx2`` or ``2xN``
-        `~astropy.units.Quantity` in units of pixels.  A ``2x2``
-        `~numpy.ndarray` or `~astropy.units.Quantity` is interpreted as
-        ``Nx2``, i.e. two rows of (x, y) coordinates.
-    a : float
-        The semimajor axis.
-    b : float
-        The semiminor axis.
-    theta : float
-        The position angle of the semimajor axis in radians
-        (counterclockwise).
-
-    Raises
-    ------
-    ValueError : `ValueError`
-        If either axis (``a`` or ``b``) is negative.
-    """
-
-    def __init__(self, positions, a, b, theta):
-        try:
-            self.a = float(a)
-            self.b = float(b)
-            self.theta = float(theta)
-        except TypeError:
-            raise TypeError("'a' and 'b' and 'theta' must be numeric, "
-                            "received {0} and {1} and {2}."
-                            .format((type(a), type(b), type(theta))))
-
-        if a < 0 or b < 0:
-            raise ValueError("'a' and 'b' must be non-negative.")
-
-        self.positions = _sanitize_pixel_positions(positions)
-
-    def area(self):
-        return math.pi * self.a * self.b
-
-    def do_photometry(self, data, error=None, pixelwise_error=True,
-                      method='exact', subpixels=5):
-
-        if method not in ('center', 'subpixel', 'exact'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        flux = do_elliptical_photometry(data, self.positions,
-                                        self.a, self.b, self.theta,
-                                        error=error,
-                                        pixelwise_error=pixelwise_error,
-                                        method=method,
-                                        subpixels=subpixels)
-        return flux
-
-    def get_fractions(self, data, method='exact', subpixels=5):
-
-        if method not in ('center', 'subpixel', 'exact'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        fractions = get_elliptical_fractions(data, self.positions,
-                                             self.a, self.b, self.theta,
-                                             method=method,
-                                             subpixels=subpixels)
-
-        return fractions
-
-    def plot(self, origin=(0, 0), source_id=None, ax=None, fill=False,
-             **kwargs):
-
-        import matplotlib.patches as mpatches
-
-        plot_positions, ax, kwargs = self._prepare_plot(
-            origin, source_id, ax, fill, **kwargs)
-
-        theta_deg = self.theta * 180. / np.pi
-        for position in plot_positions:
-            patch = mpatches.Ellipse(position, 2.*self.a, 2.*self.b,
-                                     theta_deg, **kwargs)
-            ax.add_patch(patch)
-
-
-class SkyEllipticalAnnulus(SkyAperture):
-    """
-    Elliptical annulus aperture(s), defined in sky coordinates.
-
-    Parameters
-    ----------
-    positions : `~astropy.coordinates.SkyCoord`
-        Celestial coordinates of the aperture center(s). This can be either
-        scalar coordinates or an array of coordinates.
-    a_in : `~astropy.units.Quantity`
-        The inner semimajor axis, either in angular or pixel units.
-    a_out : `~astropy.units.Quantity`
-        The outer semimajor axis, either in angular or pixel units.
-    b_out : `~astropy.units.Quantity`
-        The outer semiminor axis, either in angular or pixel units. The inner
-        semiminor axis is determined by scaling by a_in/a_out.
-    theta : `~astropy.units.Quantity`
-        The position angle of the semimajor axis (counterclockwise), either
-        in angular or pixel units.
-    """
-
-    def __init__(self, positions, a_in, a_out, b_out, theta):
-
-        if isinstance(positions, SkyCoord):
-            self.positions = positions
-        else:
-            raise TypeError("positions should be a SkyCoord instance")
-
-        assert_angle_or_pixel('a_in', a_in)
-        assert_angle_or_pixel('a_out', a_out)
-        assert_angle_or_pixel('b_out', b_out)
-        assert_angle('theta', theta)
-
-        if a_in.unit.physical_type != a_out.unit.physical_type:
-            raise ValueError("a_in and a_out should either both be angles "
-                             "or in pixels")
-
-        if a_out.unit.physical_type != b_out.unit.physical_type:
-            raise ValueError("a_out and b_out should either both be angles "
-                             "or in pixels")
-
-        self.a_in = a_in
-        self.a_out = a_out
-        self.b_out = b_out
-        self.theta = theta
-
-    def to_pixel(self, wcs):
-        """
-        Return a EllipticalAnnulus instance in pixel coordinates.
-        """
-
-        x, y = skycoord_to_pixel(self.positions, wcs,
-                                 mode=skycoord_to_pixel_mode)
-        central_pos = SkyCoord([wcs.wcs.crval], frame=self.positions.name,
-                               unit=wcs.wcs.cunit)
-        xc, yc, scale, angle = skycoord_to_pixel_scale_angle(central_pos, wcs)
-
-        if self.a_in.unit.physical_type == 'angle':
-            a_in = (scale * self.a_in).to(u.pixel).value
-            a_out = (scale * self.a_out).to(u.pixel).value
-            b_out = (scale * self.b_out).to(u.pixel).value
-        else:
-            a_in = self.a_in.value
-            a_out = self.a_out.value
-            b_out = self.b_out.value
-
-        theta = (angle + self.theta).to(u.radian).value
-
-        pixel_positions = np.array([x, y]).transpose()
-
-        return EllipticalAnnulus(pixel_positions, a_in, a_out, b_out, theta)
-
-
-class EllipticalAnnulus(PixelAperture):
-    """
-    Elliptical annulus aperture(s), defined in pixel coordinates.
-
-    Parameters
-    ----------
-    positions : tuple, list, array, or `~astropy.units.Quantity`
-        Pixel coordinates of the aperture center(s), either as a single
-        ``(x, y)`` tuple, a list of ``(x, y)`` tuples, an ``Nx2`` or
-        ``2xN`` `~numpy.ndarray`, or an ``Nx2`` or ``2xN``
-        `~astropy.units.Quantity` in units of pixels.  A ``2x2``
-        `~numpy.ndarray` or `~astropy.units.Quantity` is interpreted as
-        ``Nx2``, i.e. two rows of (x, y) coordinates.
-    a_in : float
-        The inner semimajor axis.
-    a_out : float
-        The outer semimajor axis.
-    b_out : float
-        The outer semiminor axis. (The inner semiminor axis is determined
-        by scaling by a_in/a_out.)
-    theta : float
-        The position angle of the semimajor axis in radians.
-        (counterclockwise).
-
-    Raises
-    ------
-    ValueError : `ValueError`
-        If inner semimajor axis (``a_in``) is greater than outer semimajor
-        axis (``a_out``).
-    ValueError : `ValueError`
-        If either the inner semimajor axis (``a_in``) or the outer semiminor
-        axis (``b_out``) is negative.
-    """
-
-    def __init__(self, positions, a_in, a_out, b_out, theta):
-        try:
-            self.a_in = float(a_in)
-            self.a_out = float(a_out)
-            self.b_out = float(b_out)
-            self.theta = float(theta)
-        except TypeError:
-            raise TypeError("'a_in' and 'a_out' and 'b_out' and 'theta' must "
-                            "be numeric, received {0} and {1} and {2} and "
-                            "{3}.".format((type(a_in), type(a_out),
-                                           type(b_out), type(theta))))
-
-        if not (a_out > a_in):
-            raise ValueError("'a_out' must be greater than 'a_in'")
-        if a_in < 0 or b_out < 0:
-            raise ValueError("'a_in' and 'b_out' must be non-negative")
-
-        self.b_in = a_in * b_out / a_out
-
-        self.positions = _sanitize_pixel_positions(positions)
-
-    def area(self):
-        return math.pi * (self.a_out * self.b_out - self.a_in * self.b_in)
-
-    def plot(self, origin=(0, 0), source_id=None, ax=None, fill=False,
-             **kwargs):
-
-        import matplotlib.patches as mpatches
-
-        plot_positions, ax, kwargs = self._prepare_plot(
-            origin, source_id, ax, fill, **kwargs)
-
-        theta_deg = self.theta * 180. / np.pi
-        for position in plot_positions:
-            patch_inner = mpatches.Ellipse(position, 2.*self.a_in,
-                                           2.*self.b_in, theta_deg, **kwargs)
-            patch_outer = mpatches.Ellipse(position, 2.*self.a_out,
-                                           2.*self.b_out, theta_deg, **kwargs)
-            path = _make_annulus_path(patch_inner, patch_outer)
-            patch = mpatches.PathPatch(path, **kwargs)
-            ax.add_patch(patch)
-
-    def do_photometry(self, data, error=None, pixelwise_error=True,
-                      method='exact', subpixels=5):
-
-        if method not in ('center', 'subpixel', 'exact'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        flux = do_elliptical_photometry(data, self.positions,
-                                        self.a_out, self.b_out, self.theta,
-                                        error=error,
-                                        pixelwise_error=pixelwise_error,
-                                        method=method,
-                                        subpixels=subpixels,
-                                        a_in=self.a_in)
-
-        return flux
-
-    def get_fractions(self, data, method='exact', subpixels=5):
-
-        if method not in ('center', 'subpixel', 'exact'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        fractions = get_elliptical_fractions(
-            data, self.positions, self.a_out, self.b_out, self.theta,
-            method=method, subpixels=subpixels, a_in=self.a_in)
-
-        return fractions
-
-
-class SkyRectangularAperture(SkyAperture):
-    """
-    Rectangular aperture(s), defined in sky coordinates.
-
-    Parameters
-    ----------
-    positions : `~astropy.coordinates.SkyCoord`
-        Celestial coordinates of the aperture center(s). This can be either
-        scalar coordinates or an array of coordinates.
-    w : `~astropy.units.Quantity`
-        The full width of the aperture(s) (at theta = 0, this is the "x"
-        axis), either in angular or pixel units.
-    h :  `~astropy.units.Quantity`
-        The full height of the aperture(s) (at theta = 0, this is the "y"
-        axis), either in angular or pixel units.
-    theta : `~astropy.units.Quantity`
-        The position angle of the width side in radians
-        (counterclockwise).
-    """
-
-    def __init__(self, positions, w, h, theta):
-
-        if isinstance(positions, SkyCoord):
-            self.positions = positions
-        else:
-            raise TypeError("positions should be a SkyCoord instance")
-
-        assert_angle_or_pixel('w', w)
-        assert_angle_or_pixel('h', h)
-        assert_angle('theta', theta)
-
-        if w.unit.physical_type != h.unit.physical_type:
-            raise ValueError("'w' and 'h' should either both be angles or "
-                             "in pixels")
-
-        self.w = w
-        self.h = h
-        self.theta = theta
-
-    def to_pixel(self, wcs):
-        """
-        Return a RectangularAperture instance in pixel coordinates.
-        """
-
-        x, y = skycoord_to_pixel(self.positions, wcs,
-                                 mode=skycoord_to_pixel_mode)
-        central_pos = SkyCoord([wcs.wcs.crval], frame=self.positions.name,
-                               unit=wcs.wcs.cunit)
-        xc, yc, scale, angle = skycoord_to_pixel_scale_angle(central_pos, wcs)
-
-        if self.w.unit.physical_type == 'angle':
-            w = (scale * self.w).to(u.pixel).value
-            h = (scale * self.h).to(u.pixel).value
-        else:  # pixel
-            w = self.w.value
-            h = self.h.value
-
-        theta = (angle + self.theta).to(u.radian).value
-
-        pixel_positions = np.array([x, y]).transpose()
-
-        return RectangularAperture(pixel_positions, w, h, theta)
-
-
-class RectangularAperture(PixelAperture):
-    """
-    Rectangular aperture(s), defined in pixel coordinates.
-
-    Parameters
-    ----------
-    positions : tuple, or list, or array
-        Pixel coordinates of the aperture center(s), either as a single
-        ``(x, y)`` tuple, a list of ``(x, y)`` tuples, an ``Nx2`` or
-        ``2xN`` `~numpy.ndarray`, or an ``Nx2`` or ``2xN``
-        `~astropy.units.Quantity` in units of pixels.  A ``2x2``
-        `~numpy.ndarray` or `~astropy.units.Quantity` is interpreted as
-        ``Nx2``, i.e. two rows of (x, y) coordinates.
-    w : float
-        The full width of the aperture (at theta = 0, this is the "x" axis).
-    h : float
-        The full height of the aperture (at theta = 0, this is the "y" axis).
-    theta : float
-        The position angle of the width side in radians
-        (counterclockwise).
-
-    Raises
-    ------
-    ValueError : `ValueError`
-        If either width (``w``) or height (``h``) is negative.
-    """
-
-    def __init__(self, positions, w, h, theta):
-        try:
-            self.w = float(w)
-            self.h = float(h)
-            self.theta = float(theta)
-        except TypeError:
-            raise TypeError("'w' and 'h' and 'theta' must "
-                            "be numeric, received {0} and {1} and {2}."
-                            .format((type(w), type(h), type(theta))))
-        if w < 0 or h < 0:
-            raise ValueError("'w' and 'h' must be nonnegative.")
-
-        self.positions = _sanitize_pixel_positions(positions)
-
-    def area(self):
-        return self.w * self.h
-
-    def plot(self, origin=(0, 0), source_id=None, ax=None, fill=False,
-             **kwargs):
-
-        import matplotlib.patches as mpatches
-
-        plot_positions, ax, kwargs = self._prepare_plot(
-            origin, source_id, ax, fill, **kwargs)
-
-        hw = self.w / 2.
-        hh = self.h / 2.
-        sint = math.sin(self.theta)
-        cost = math.cos(self.theta)
-        dx = (hh * sint) - (hw * cost)
-        dy = -(hh * cost) - (hw * sint)
-        plot_positions = plot_positions + np.array([dx, dy])
-        theta_deg = self.theta * 180. / np.pi
-        for position in plot_positions:
-            patch = mpatches.Rectangle(position, self.w, self.h, theta_deg,
-                                       **kwargs)
-            ax.add_patch(patch)
-
-    def do_photometry(self, data, error=None, pixelwise_error=True,
-                      method='subpixel', subpixels=5):
-
-        if method == 'exact':
-            warnings.warn("'exact' method is not implemented, defaults to "
-                          "'subpixel' method and subpixels=32 instead",
-                          AstropyUserWarning)
-            method = 'subpixel'
-            subpixels = 32
-
-        elif method not in ('center', 'subpixel'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        flux = do_rectangular_photometry(data, self.positions,
-                                         self.w, self.h, self.theta,
-                                         error=error,
-                                         pixelwise_error=pixelwise_error,
-                                         method=method,
-                                         subpixels=subpixels)
-        return flux
-
-    def get_fractions(self, data, method='exact', subpixels=5):
-
-        if method == 'exact':
-            warnings.warn("'exact' method is not implemented, defaults to "
-                          "'subpixel' method and subpixels=32 instead",
-                          AstropyUserWarning)
-            method = 'subpixel'
-            subpixels = 32
-
-        elif method not in ('center', 'subpixel'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        fractions = get_rectangular_fractions(data, self.positions,
-                                              self.w, self.h, self.theta,
-                                              method=method,
-                                              subpixels=subpixels)
-
-        return fractions
-
-
-class SkyRectangularAnnulus(SkyAperture):
-    """
-    Rectangular annulus aperture(s), defined in sky coordinates.
-
-    Parameters
-    ----------
-    positions : `~astropy.coordinates.SkyCoord`
-        Celestial coordinates of the aperture center(s). This can be either
-        scalar coordinates or an array of coordinates.
-    w_in : `~astropy.units.Quantity`
-        The inner full width of the aperture(s), either in angular or pixel
-        units.
-    w_out : `~astropy.units.Quantity`
-        The outer full width of the aperture(s), either in angular or pixel
-        units.
-    h_out : `~astropy.units.Quantity`
-        The outer full height of the aperture(s), either in angular or pixel
-        units. (The inner full height is determined by scaling by
-        w_in/w_out.)
-    theta : `~astropy.units.Quantity`
-        The position angle of the semimajor axis (counterclockwise), either
-        in angular or pixel units.
-    """
-
-    def __init__(self, positions, w_in, w_out, h_out, theta):
-
-        if isinstance(positions, SkyCoord):
-            self.positions = positions
-        else:
-            raise TypeError("positions should be a SkyCoord instance")
-        assert_angle_or_pixel('w_in', w_in)
-        assert_angle_or_pixel('w_out', w_out)
-        assert_angle_or_pixel('h_out', h_out)
-        assert_angle('theta', theta)
-
-        if w_in.unit.physical_type != w_out.unit.physical_type:
-            raise ValueError("w_in and w_out should either both be angles or "
-                             "in pixels")
-
-        if w_out.unit.physical_type != h_out.unit.physical_type:
-            raise ValueError("w_out and h_out should either both be angles "
-                             "or in pixels")
-
-        self.w_in = w_in
-        self.w_out = w_out
-        self.h_out = h_out
-        self.theta = theta
-
-    def to_pixel(self, wcs):
-        """
-        Return a EllipticalAnnulus instance in pixel coordinates.
-        """
-
-        x, y = skycoord_to_pixel(self.positions, wcs,
-                                 mode=skycoord_to_pixel_mode)
-        central_pos = SkyCoord([wcs.wcs.crval], frame=self.positions.name,
-                               unit=wcs.wcs.cunit)
-        xc, yc, scale, angle = skycoord_to_pixel_scale_angle(central_pos, wcs)
-
-        if self.w_in.unit.physical_type == 'angle':
-            w_in = (scale * self.w_in).to(u.pixel).value
-            w_out = (scale * self.w_out).to(u.pixel).value
-            h_out = (scale * self.h_out).to(u.pixel).value
-        else:
-            w_in = self.w_in.value
-            w_out = self.w_out.value
-            h_out = self.h_out.value
-
-        theta = (angle + self.theta).to(u.radian).value
-
-        pixel_positions = np.array([x, y]).transpose()
-
-        return RectangularAnnulus(pixel_positions, w_in, w_out, h_out, theta)
-
-
-class RectangularAnnulus(PixelAperture):
-    """
-    Rectangular annulus aperture(s), defined in pixel coordinates.
-
-    Parameters
-    ----------
-    positions : tuple, list, array, or `~astropy.units.Quantity`
-        Pixel coordinates of the aperture center(s), either as a single
-        ``(x, y)`` tuple, a list of ``(x, y)`` tuples, an ``Nx2`` or
-        ``2xN`` `~numpy.ndarray`, or an ``Nx2`` or ``2xN``
-        `~astropy.units.Quantity` in units of pixels.  A ``2x2``
-        `~numpy.ndarray` or `~astropy.units.Quantity` is interpreted as
-        ``Nx2``, i.e. two rows of (x, y) coordinates.
-    w_in : float
-        The inner full width of the aperture.
-    w_out : float
-        The outer full width of the aperture.
-    h_out : float
-        The outer full height of the aperture. (The inner full height is
-        determined by scaling by w_in/w_out.)
-    theta : float
-        The position angle of the width side in radians.
-        (counterclockwise).
-
-    Raises
-    ------
-    ValueError : `ValueError`
-        If inner width (``w_in``) is greater than outer width (``w_out``).
-    ValueError : `ValueError`
-        If either the inner width (``w_in``) or the outer height (``h_out``)
-        is negative.
-    """
-
-    def __init__(self, positions, w_in, w_out, h_out, theta):
-        try:
-            self.w_in = float(w_in)
-            self.w_out = float(w_out)
-            self.h_out = float(h_out)
-            self.theta = float(theta)
-        except TypeError:
-            raise TypeError("'w_in' and 'w_out' and 'h_out' and 'theta' must "
-                            "be numeric, received {0} and {1} and {2} and "
-                            "{3}.".format((type(w_in), type(w_out),
-                                           type(h_out), type(theta))))
-
-        if not (w_out > w_in):
-            raise ValueError("'w_out' must be greater than 'w_in'")
-        if w_in < 0 or h_out < 0:
-            raise ValueError("'w_in' and 'h_out' must be non-negative")
-
-        self.h_in = w_in * h_out / w_out
-
-        self.positions = _sanitize_pixel_positions(positions)
-
-    def area(self):
-        return self.w_out * self.h_out - self.w_in * self.h_in
-
-    def plot(self, origin=(0, 0), source_id=None, ax=None, fill=False,
-             **kwargs):
-
-        import matplotlib.patches as mpatches
-
-        plot_positions, ax, kwargs = self._prepare_plot(
-            origin, source_id, ax, fill, **kwargs)
-
-        sint = math.sin(self.theta)
-        cost = math.cos(self.theta)
-        theta_deg = self.theta * 180. / np.pi
-
-        hw_inner = self.w_in / 2.
-        hh_inner = self.h_in / 2.
-        dx_inner = (hh_inner * sint) - (hw_inner * cost)
-        dy_inner = -(hh_inner * cost) - (hw_inner * sint)
-        positions_inner = plot_positions + np.array([dx_inner, dy_inner])
-        hw_outer = self.w_out / 2.
-        hh_outer = self.h_out / 2.
-        dx_outer = (hh_outer * sint) - (hw_outer * cost)
-        dy_outer = -(hh_outer * cost) - (hw_outer * sint)
-        positions_outer = plot_positions + np.array([dx_outer, dy_outer])
-
-        for i, position_inner in enumerate(positions_inner):
-            patch_inner = mpatches.Rectangle(position_inner,
-                                             self.w_in, self.h_in,
-                                             theta_deg, **kwargs)
-            patch_outer = mpatches.Rectangle(positions_outer[i],
-                                             self.w_out, self.h_out,
-                                             theta_deg, **kwargs)
-            path = _make_annulus_path(patch_inner, patch_outer)
-            patch = mpatches.PathPatch(path, **kwargs)
-            ax.add_patch(patch)
-
-    def do_photometry(self, data, error=None, pixelwise_error=True,
-                      method='subpixel', subpixels=5):
-
-        if method == 'exact':
-            warnings.warn("'exact' method is not implemented, defaults to "
-                          "'subpixel' instead", AstropyUserWarning)
-            method = 'subpixel'
-
-        elif method not in ('center', 'subpixel'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        flux = do_rectangular_photometry(data, self.positions,
-                                         self.w_out, self.h_out, self.theta,
-                                         error=error,
-                                         pixelwise_error=pixelwise_error,
-                                         method=method, subpixels=subpixels,
-                                         w_in=self.w_in)
-
-        return flux
-
-    def get_fractions(self, data, method='exact', subpixels=5):
-
-        if method == 'exact':
-            warnings.warn("'exact' method is not implemented, defaults to "
-                          "'subpixel' method and subpixels=32 instead",
-                          AstropyUserWarning)
-            method = 'subpixel'
-            subpixels = 32
-
-        elif method not in ('center', 'subpixel'):
-            raise ValueError('{0} method not supported for aperture class '
-                             '{1}'.format(method, self.__class__.__name__))
-
-        fractions = get_rectangular_fractions(data, self.positions,
-                                              self.w_out, self.h_out,
-                                              self.theta,
-                                              method=method,
-                                              subpixels=subpixels,
-                                              w_in=self.w_in)
-
-        return fractions
 
 
 def _prepare_photometry_input(data, unit, wcs, mask, error, pixelwise_error):
